isolate some common function used by the logging in a separate header file -- print metadata in human readable format when the logs are requested by the regulator
