+++ controller/source/default_policy.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <string>
-#include <bitset>
 #include <sstream>
 #include <unordered_map>
